Changed simplify to work on tables in place.
